requirements.txt
numpy==1.24.3
urllib3==1.26.15
pytest==7.2.2
allure-pytest==2.13.2
pymysql==1.0.3
python-dotenv==1.0.0
mysqlclient==2.1.1
webdriver-manager==4.0.1
selenium==4.12.0
pandas==2.0.3
openpyxl==3.1.2
pytest-xdist==3.3.1
pytest-rerunfailures==11.1.2